Batching to handle variable number of chunks, and final chunk of variable size. Introduce JPEG compression of video frames, with 50% (lossy) compression we have gone from 2MB of audio/visual data to 100kb of audio/visual data per second. This includes uncompressed audio data, video data is about 40+ times smaller. AWESOME!!!

+++ MyFirstApplication/Batch.h
@@ -10,6 +10,6 @@
 #import "PipelineProcessor.h"
 
 @interface Batch : PipelineProcessor
-- (id)initWithOutputSession:(id<NewPacketDelegate>)outputSession chunkSize:(uint)chunkSize numChunks:(uint)numChunks andNumChunksThreshold:(uint)numChunksThreshold andTimeoutSeconds:(double)timeoutSeconds;
+- (id)initWithOutputSession:(id<NewPacketDelegate>)outputSession chunkSize:(uint)chunkSize numChunks:(uint)numChunks andNumChunksThreshold:(float)numChunksThreshold andTimeoutSeconds:(double)timeoutSeconds;
 - (void)onNewPacket:(ByteBuffer*)packet fromProtocol:(ProtocolType)protocol;
 @end
